+++ CHANGELOG.md
@@ -5,9 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.1.1] - 2018-8-10
 ### Added
 - `conjurctl export` now includes the account list to support migration
-
 - Update puma to a version that understands how to handle having ipv6 disabled
 - Update puma worker timeout to allow longer requests to finish (from 1 minute to 10 minutes)